requirements.txt
Pillow
cairosvg
streamlit